% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_ICP}
\alias{age_ICP}
\title{Absolute FT ages}
\usage{
age_ICP(x, spotsize = 40, zeta, ...)
}
\arguments{
\item{x}{\code{data.frame} containing counts of spontaneous tracks,
U concentrations, and the uncertainties of U concentrations}

\item{spotsize}{the laser ablation spot size (diameter in micrometer)}

\item{zeta}{two-element vector with the zeta-factor and its standard error.}

\item{...}{additional arguments}
}
\value{
\code{list} with the single grain ages and their analytical uncertainties,
the pooled age, and the chisq statistics
}
\description{
Calculation of the ages from LA-ICP-MS using zeta calibration
}
\details{
The fission track age using the zeta calibration approach
(Hasebe et al. 2009) is given by
\deqn{t = \frac{1}{\lambda} \ln \left(1+\frac{1}{2} \lambda \zeta_{icp} \frac{N)s}{A_s [ ^{238}U/^xX]} \right)}
where \eqn{\lambda}{lambda} is the 238U decay constant,
\eqn{\zeta_{icp}}{zeta_icp} is the zeta calibration factor,
\eqn{N_s}{Ns} is the count of spontaneous fission tracks,
\eqn{A_s}{As} is the size of the ablated area, and
\eqn{[ ^{238}U/^xX]}{238U} either stands for the
\eqn{^{238}}{238}U-concentration (in ppm) or for the U/Ca (for apatite) or
U/Si (for zircon) ratio measurement.
}
\examples{
data("sample")
age_ICP(sample, spotsize = 40, zeta = c(0.1188, 0.0119))
}
\references{
Hasebe, N., Carter, A., Hurford, A.J., Arai, S., 2009. The effect of chemical etching on LA-ICP-MS analysis in determining uranium concentration for fission-track chronometry. Geol. Soc. Lond. Spec. Publ. 324 (1), 37--46. \doi{10.1144/SP324.3}

Vermeesch, Pieter, 2017. Statistics for LA-ICP-MS based fission track dating. Chem. Geol. 456, 19--27. \doi{10.1016/j.chemgeo.2017.03.002}
}
