% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as.fissiontracks}
\alias{as.fissiontracks}
\title{Export to IsoplotR}
\usage{
as.fissiontracks(x, spotsize = 40, zeta, ages = FALSE)
}
\arguments{
\item{x}{a data.frame of spontaneous and fission track counts and the U-concentration}

\item{spotsize}{the laser ablation spot size}

\item{zeta}{the zeta calibration constant extracted from the input data}

\item{ages}{logical. \code{FALSE} if \code{x} represents unprocessed data
(the default), or \code{TRUE} if processed ages}
}
\value{
An object of class \code{fissiontracks}
}
\description{
transform into a \code{"fissiontracks"} object
}
\examples{
data(standard)
session.zeta <- zeta_ICP(standard, spotsize = 40, mineral = "apatite", standard = "Dur")
data("sample")
as.fissiontracks(sample, spotsize = 40, zeta = session.zeta$zeta)
}
