% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chi_stats}
\alias{chi_stats}
\title{Chisq stats}
\usage{
chi_stats(t, st, na.rm = TRUE)
}
\arguments{
\item{t}{numeric. ages}

\item{st}{numeric. analytical uncertainties}

\item{na.rm}{logical. whether NA values should be removed (TRUE) or not (FALSE)}
}
\description{
Chisq stats
}
