% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ages}
\alias{get_ages}
\title{Single-grain FT age FT ages}
\usage{
get_ages(x, rhos, zeta)
}
\arguments{
\item{x}{\code{"dataframe"} with count of spontaneous fission-tracks.
containing \code{U} U-concentration, given as \eqn{^{238}}{238}U/\eqn{^{29}}{29}Ca based on cps (for zircon), \eqn{^{238}}{238}U/\eqn{^{43}}{43}Ca based on cps (apatite), or \eqn{^{238}}{238}U in ppm
and \code{sU} standard deviation of \code{U}}

\item{rhos}{numeric. Spontaneous fission track density}

\item{zeta}{numeric. Calibration factor based on FCT2 fission-track age standard}
}
\description{
blabla bla
}
\note{
You can either use U ratio cps, or U ppm. Whatever you choose for zeta calculation, you must use it for age calculation, too.
}
