% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_prob_chisq}
\alias{get_prob_chisq}
\title{Probability of chi-squared}
\usage{
get_prob_chisq(chisq, n)
}
\arguments{
\item{chisq}{Chi-squared statistic}

\item{n}{Number of grains}
}
\description{
Probability of chi-squared
}
\examples{
get_prob_chisq(21.0991, 47)
}
