% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_rhos}
\alias{get_rhos}
\title{spontaneous fission-track density}
\usage{
get_rhos(Ns, area)
}
\arguments{
\item{Ns}{spontaneous fission track count}

\item{area}{Area}
}
\description{
spontaneous fission-track density
}
