% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_zeta_error}
\alias{get_zeta_error}
\title{Zeta error}
\usage{
get_zeta_error(x, zeta)
}
\arguments{
\item{x}{\code{"dataframe"} with \code{Ns}, \code{st}, \code{t}, \code{U}, \code{sU}}

\item{zeta}{calibration factor based on FCT2 fission-track age standard}
}
\description{
Zeta error
}
\note{
see data("sample") for details on \code{x}
}
