% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getkde}
\alias{getkde}
\title{Adaptive kernel density estimates}
\usage{
getkde(
  x,
  from = NA,
  to = NA,
  bw = NA,
  adaptive = TRUE,
  log = FALSE,
  n = 512,
  weights = NULL,
  ...
)
}
\arguments{
\item{x}{numeric vector. Ages}

\item{from}{numeric. minimum age of the time axis. If NULL, this is set automatically}

\item{to}{numeric. maximum age of the time axis. If NULL, this is set automatically}

\item{bw}{numeric. the bandwidth of the KDE. If NULL, bw will be calculated
automatically using the algorithm by Botev et al. (2010).}

\item{adaptive}{logical flag controlling if the adaptive KDE modifier of Abramson (1982) is used}

\item{log}{logical. transform the ages to a log scale if TRUE}

\item{n}{integer. horizontal resolution (i.e., the number of segments) of the density estimate.}

\item{weights}{numeric vector of non-negative observation weights, hence of
same length as \code{x}. The default \code{NULL} is equivalent to
\code{weights = rep(1/nx, nx)} where \code{nx} is the length of (the finite entries of)
\code{x[]}. If \code{na.rm = TRUE} and there are \code{NA}'s in \code{x}, they and the
corresponding weights are removed before computations.
In that case, when the original weights have summed to one, they are
re-scaled to keep doing so. Will be ignored is \code{adaptive} is \code{TRUE}.}

\item{...}{optional arguments to be passed on to R's density function.}
}
\description{
Modification of \code{\link[ggplot2:geom_density]{ggplot2::stat_density()}} for kernel density estimates using a combination of the
Botev (2010) bandwidth selector and the Abramson (1982) adaptive kernel
bandwidth modifier.
}
