% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pooled_age}
\alias{pooled_age}
\title{Pooled age}
\usage{
pooled_age(x, r, zeta)
}
\arguments{
\item{x}{dataset}

\item{r}{Radius}

\item{zeta}{two-element vector with the zeta-factor and its standard error.}
}
\description{
Pooled age
}
