% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{sample}
\alias{sample}
\title{Example dataset}
\format{
An object of class \code{"tibble"}
}
\usage{
data('sample')
}
\description{
Taís's FCT measurements
}
\examples{
data("sample")
head("sample")
}
\keyword{datasets}
