% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zeta_ICP}
\alias{zeta_ICP}
\title{Zeta calibration coefficient for LA-ICP-MS fission track dating}
\usage{
zeta_ICP(
  x,
  spotsize = 40,
  mineral = c("zircon", "apatite"),
  standard = c("Dur", "FCT"),
  tst,
  lambda = pkg.env$lambda[1],
  g = pkg.env$g,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} containing counts of spontaneous tracks, U concentrations, and the uncertainties of U concentrations}

\item{spotsize}{the laser ablation spot size (diameter in micrometer)}

\item{mineral}{Mineral of the standard, one of 'zircon' or 'apatite'}

\item{standard}{Name of the standard, one of 'Dur' or 'FCT'}

\item{tst}{(optional) two-column vector giving the true age and its standard error. overrides mineral and standard}

\item{lambda}{(optional) decay constant}

\item{g}{(optional) geometry factor (default: 0.5)}

\item{...}{additional arguments}
}
\value{
list containing: two-element vector with the session zeta and its standard error
}
\description{
Calibrate the ICP-session zeta factor using the number of spontaneous tracks,
the area over which these were counted and one ore more U/Ca, U/Si or
U-concentration measurements (in ppm) and their analytical uncertainties.
}
\examples{
data("standard")
zeta_ICP(standard, spotsize = 40, mineral = "apatite", standard = "Dur")
}
\references{
Hasebe, N., Carter, A., Hurford, A.J., Arai, S., 2009. The effect of chemical etching on LA-ICP-MS analysis in determining uranium concentration for fission-track chronometry. Geol. Soc. Lond. Spec. Publ. 324 (1), 37--46. \doi{10.1144/SP324.3}
}
