% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{aKDE}
\alias{aKDE}
\alias{stat_aKDE}
\alias{geom_aKDE}
\title{Adaptive kernel density estimates}
\source{
Algorithm for adaptive kernel is modified from \link{IsoplotR}. The
algorithm for the optimal kernel bandwidth is from \code{\link[provenance:botev]{provenance::botev()}}.
}
\usage{
stat_aKDE(
  mapping = NULL,
  data = NULL,
  stat = "DensityAdaptive",
  geom = "area",
  position = "stack",
  ...,
  from = NA,
  to = NA,
  bw = NA,
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  adaptive = TRUE,
  na.rm = FALSE,
  bounds = c(-Inf, Inf),
  show.legend = NA,
  orientation = NA,
  inherit.aes = TRUE
)

geom_aKDE(
  mapping = NULL,
  data = NULL,
  stat = "DensityAdaptive",
  position = "identity",
  ...,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom, stat}{Use to override the default connection between
\code{geom_density()} and \code{stat_density()}.}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{from, to}{the left and right-most points of the grid at which the density is to be estimated}

\item{bw}{the bandwidth of the KDE. If \code{NULL}, \code{bw} will be calculated automatically using the algorithm by Botev et al. (2010).}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, \code{adjust = 1/2} means use half of the default bandwidth.}

\item{kernel}{Kernel. See list of available kernels in \code{\link[=density]{density()}}.}

\item{n}{number of equally spaced points at which the density is to be
estimated, should be a power of two, see \code{\link[=density]{density()}} for
details}

\item{adaptive}{logical flag controlling if the adaptive KDE modifier of Abramson (1982) is used}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{bounds}{Known lower and upper bounds for estimated data. Default
\code{c(-Inf, Inf)} means that there are no (finite) bounds. If any bound is
finite, boundary effect of default density estimation will be corrected by
reflecting tails outside \code{bounds} around their closest edge. Data points
outside of bounds are removed with a warning.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{outline.type}{Type of the outline of the area; \code{"both"} draws both the
upper and lower lines, \code{"upper"}/\code{"lower"} draws the respective lines only.
\code{"full"} draws a closed polygon around the area.}
}
\description{
Modification of \code{\link[ggplot2:geom_density]{ggplot2::stat_density()}} for kernel density estimates using a combination of the
Botev (2010) bandwidth selector and the Abramson (1982) adaptive kernel
bandwidth modifier.
}
\examples{
data("sample")
example <- age_ICP(sample, zeta = c(0.1188, 0.0119))
# IsoplotR::kde(example$ages$t)
ggplot2::ggplot(data = example$ages, mapping = ggplot2::aes(x = t)) +
  stat_aKDE(adaptive = TRUE) +
  stat_aKDE(adaptive = FALSE, color = "red", fill = NA)

ggplot2::ggplot(
data = example$ages,
mapping = ggplot2::aes(x = t, weight = t / st)) +
  geom_aKDE(
  ggplot2::aes(y = ggplot2::after_stat(scaled)),
  kernel = "epanechnikov", fill = "steelblue", alpha = .75) +
  ggplot2::geom_histogram(
  ggplot2::aes(y = ggplot2::after_stat(ncount)),
  color = "grey", fill = "grey", alpha = .5) +
  ggplot2::geom_rug(alpha = 0.5)
}
