% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chi_stats}
\alias{chi_stats}
\title{Chisq stats}
\usage{
chi_stats(t, st, na.rm = TRUE)
}
\description{
Chisq stats
}
\examples{
data("sample")
ages <- get_ages(sample, zeta = c(0.1188, 0.0119))
chi_stats(ages$t[1:25], ages$st[1:25])
}
