% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_chisq}
\alias{get_chisq}
\title{Chi-squared statistic}
\usage{
get_chisq(t, st, na.rm = TRUE)
}
\arguments{
\item{t}{ages}

\item{st}{analytical uncertainties}

\item{na.rm}{logical. wether NA values should be removed (TRUE) or not (FALSE)}
}
\description{
Chi-squared statistic
}
