% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_prob_chisq}
\alias{get_prob_chisq}
\title{Probability of chi-squared}
\usage{
get_prob_chisq(chisq, n)
}
\arguments{
\item{chisq}{numeric. Chi-squared statistic}

\item{n}{integer. Number of grains}
}
\description{
Probability of chi-squared
}
