% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard}
\alias{standard}
\title{Example zeta dataset}
\format{
An object of class \code{"tibble"}
}
\usage{
data('standard')
}
\description{
Taís's DUR measurements
}
\examples{
data("standard")
head("standard")
}
\keyword{datasets}
